+++ src/pages/QSpace/QSpaceDetail/components/DetailsHeader/DetailsHead.styles.ts
