+++ src/pages/QSpace/QSpaceDetail/styles.ts
