+++ src/pages/QSpace/QSpaceMain/components/GroupStateCheckBox/GroupStateCheckBox.styles.ts
