+++ src/pages/QSpace/QSpaceMain/styles.ts
