+++ src/pages/QSpace/QSpaceMember/components/ReSignButton/ReSignButton.styles.ts
