--- src/pages/QSpaceMember/components/ReSignButton/ReSignButton.tsx
@@ -1,16 +1,0 @@
-import { StyledButton } from '@/pages/QSpaceMember/components/ReSignButton/ReSignButton.styles';
-
-interface ReSignButtonProps {
-  onClick?: () => void;
-  disabled?: boolean;
-}
-
-const ReSignButton = ({ onClick, disabled = false }: ReSignButtonProps) => {
-  return (
-    <StyledButton onClick={onClick} disabled={disabled}>
-      내보내기
-    </StyledButton>
-  );
-};
-
-export default ReSignButton;
