+++ src/router/index.tsx
@@ -3,13 +3,13 @@ import Login from '@/pages/Login';
 import ChatList from '@/pages/ChatList';
 import Main from '@/pages/Main';
 import MyPage from '@/pages/MyPage';
-import QSpaceMainPage from '@/pages/QSpaceMain';
-import CategorySelectPage from '@/pages/QSpacePost/CategorySelectPage';
+import QSpaceMainPage from '@/pages/QSpace/QSpaceMain';
+import CategorySelectPage from '@/pages/QSpace/QSpacePost/CategorySelectPage';
 import NotificationPage from '@/pages/Alarm';
-import PostGroupPage from '@/pages/QSpacePost/PostGroupPage';
+import PostGroupPage from '@/pages/QSpace/QSpacePost/PostGroupPage';
 import { createBrowserRouter } from 'react-router-dom';
 import ChatRoom from '@/pages/ChatRoom';
-import QSpaceDetailPage from '@/pages/QSpaceDetail';
+import QSpaceDetailPage from '@/pages/QSpace/QSpaceDetail';
 import QuestionPage from '@/pages/Question';
 import ProfilePage from '@/pages/Profile';
 import HobbySelectPage from '@/pages/HobbySelect';
@@ -20,7 +20,7 @@ import { RegisterPage } from '@/pages/Register';
 import { PasswordRecoveryPage } from '@/pages/PasswordRecovery';
 import { ResetPasswordPage } from '@/pages/ResetPassword';
 import { IDRecoveryPage } from '@/pages/IDRecovery';
-import QSpaceMemberPage from '@/pages/QSpaceMember';
+import QSpaceMemberPage from '@/pages/QSpace/QSpaceMember';
 import EventPage from '@/pages/Event';
 import SearchResultPage from '@/pages/Search';
 import FollowerFollowingPage from '@/pages/Following';
